+++ packages/backend/src/modules/tvl2/repositories/AmountRepository.ts
@@ -6,10 +6,6 @@ import {
   CheckConvention,
 } from '../../../peripherals/database/BaseRepository'
 import { Database } from '../../../peripherals/database/Database'
-import {
-  AmountConfigurationRecord,
-  AmountConfigurationRow,
-} from './AmountConfigurationRepository'
 
 export interface AmountRow {
   configuration_id: number
@@ -23,6 +19,34 @@ export interface AmountRecord {
   amount: bigint
 }
 
+export interface AmountConfigurationRow {
+  id: number
+  indexer_id: string
+  project_id: string
+  chain: string
+  address: string
+  escrow_address: string | null
+  origin: string
+  type: string
+  include_in_total: boolean
+  since_timestamp_inclusive: Date
+  until_timestamp_inclusive: Date | null
+}
+
+export interface AmountConfigurationRecord {
+  id: number
+  indexerId: string
+  projectId: ProjectId
+  chain: string
+  address: EthereumAddress | 'native'
+  escrowAddress?: EthereumAddress
+  origin: 'canonical' | 'external' | 'native'
+  type: 'totalSupply' | 'circulatingSupply' | 'escrow'
+  includeInTotal: boolean
+  sinceTimestampInclusive: UnixTime
+  untilTimestampInclusive?: UnixTime
+}
+
 export interface AmountWithMetadata
   extends AmountRecord,
     Omit<AmountConfigurationRecord, 'id'> {}
@@ -33,6 +57,7 @@ export class AmountRepository extends BaseRepository {
     this.autoWrap<CheckConvention<AmountRepository>>(this)
   }
 
+  // #region amounts
   async addMany(records: AmountRecord[]) {
     const rows: AmountRow[] = records.map(toRow)
     const knex = await this.knex()
@@ -83,6 +108,52 @@ export class AmountRepository extends BaseRepository {
       .delete()
   }
 
+  // #endregion
+
+  // #region configurations
+
+  async addManyConfigurations(
+    records: Omit<AmountConfigurationRecord, 'id'>[],
+  ): Promise<number[]> {
+    const rows: Omit<AmountConfigurationRow, 'id'>[] =
+      records.map(toConfigurationRow)
+
+    const knex = await this.knex()
+    const inserted = (await knex
+      .batchInsert('amounts_configurations', rows, 5_000)
+      // @ts-expect-error knex types are wrong
+      .returning('id')) as unknown as { id: number }[]
+
+    return inserted.map((row) => row.id)
+  }
+
+  async getConfigurationByIndexerId(
+    indexerId: string,
+  ): Promise<AmountConfigurationRecord[]> {
+    const knex = await this.knex()
+    const rows = await knex('amounts_configurations').where(
+      'indexer_id',
+      indexerId,
+    )
+    return rows.map(toConfigurationRecord)
+  }
+
+  async setUntilTimestampInclusive(
+    id: number,
+    untilTimestampInclusive: UnixTime | undefined,
+  ): Promise<number> {
+    const knex = await this.knex()
+    return await knex('amounts_configurations')
+      .where({ id })
+      .update({
+        until_timestamp_inclusive: untilTimestampInclusive
+          ? untilTimestampInclusive.toDate()
+          : null,
+      })
+  }
+
+  // #endregion
+
   // #region methods used only in tests
 
   async getAll(): Promise<AmountRecord[]> {
@@ -94,6 +165,17 @@ export class AmountRepository extends BaseRepository {
   async deleteAll() {
     const knex = await this.knex()
     return knex('amounts').delete()
+  }
+
+  async getAllConfigurations(): Promise<AmountConfigurationRecord[]> {
+    const knex = await this.knex()
+    const rows = await knex('amounts_configurations')
+    return rows.map(toConfigurationRecord)
+  }
+
+  async deleteAllConfigurations() {
+    const knex = await this.knex()
+    return knex('amounts_configurations').delete()
   }
 
   // #endregion
@@ -142,4 +224,49 @@ function toRecordWithMetadata(
         }
       : {}),
   }
+}
+
+function toConfigurationRow(
+  row: Omit<AmountConfigurationRecord, 'id'>,
+): Omit<AmountConfigurationRow, 'id'> {
+  return {
+    indexer_id: row.indexerId,
+    project_id: row.projectId.toString(),
+    chain: row.chain,
+    address: row.address === 'native' ? 'native' : row.address.toString(),
+    escrow_address: row.escrowAddress?.toString() ?? null,
+    origin: row.origin,
+    type: row.type,
+    include_in_total: row.includeInTotal,
+    since_timestamp_inclusive: row.sinceTimestampInclusive.toDate(),
+    until_timestamp_inclusive: row.untilTimestampInclusive?.toDate() ?? null,
+  }
+}
+
+function toConfigurationRecord(
+  row: AmountConfigurationRow,
+): AmountConfigurationRecord {
+  const r: Omit<AmountConfigurationRecord, 'escrowAddress'> = {
+    id: row.id,
+    indexerId: row.indexer_id,
+    projectId: ProjectId(row.project_id),
+    chain: row.chain,
+    address: row.address === 'native' ? 'native' : EthereumAddress(row.address),
+    ...(row.escrow_address
+      ? { escrowAddress: EthereumAddress(row.escrow_address) }
+      : {}),
+    origin: row.origin as 'canonical' | 'external' | 'native',
+    type: row.type as 'totalSupply' | 'circulatingSupply' | 'escrow',
+    includeInTotal: row.include_in_total,
+    sinceTimestampInclusive: UnixTime.fromDate(row.since_timestamp_inclusive),
+    ...(row.until_timestamp_inclusive
+      ? {
+          untilTimestampInclusive: UnixTime.fromDate(
+            row.until_timestamp_inclusive,
+          ),
+        }
+      : {}),
+  }
+
+  return r
 }
